+++ source/_internals/typescriptTestUtils.ts
@@ -1,0 +1,2 @@
+export const mixedList = [1, 'foo', 3, 'bar'];
+export const mixedListConst = [1, 'foo', 3, 'bar'] as const